+++ src/org/loklak/api/server/push/NetmonPushServlet.java
@@ -20,8 +20,6 @@ package org.loklak.api.server.push;
 
 import org.loklak.api.client.ClientConnection;
 import org.loklak.api.server.RemoteAccess;
-import org.loklak.api.server.helper.PushReport;
-import org.loklak.api.server.helper.PushServletHelper;
 import org.loklak.data.DAO;
 import org.loklak.harvester.JsonFieldConverter;
 import org.loklak.harvester.SourceType;
