+++ src/org/loklak/api/server/push/NodelistPushServlet.java
@@ -23,10 +23,7 @@ import com.github.fge.jsonschema.core.report.ProcessingReport;
 import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.common.xcontent.json.JsonXContent;
 import org.loklak.api.client.ClientConnection;
-import org.loklak.api.server.PushServlet;
 import org.loklak.api.server.RemoteAccess;
-import org.loklak.api.server.helper.PushReport;
-import org.loklak.api.server.helper.PushServletHelper;
 import org.loklak.data.DAO;
 import org.loklak.harvester.JsonFieldConverter;
 import org.loklak.harvester.JsonValidator;
